fix(etc): read meta ui data from etc/data
